# proto/goddess/goddess/api/v1/self_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: goddess/api/v1/self.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSelfServiceChangeAvatar = "/goddess.api.v1.SelfService/ChangeAvatar"
const OperationSelfServiceChangeEmail = "/goddess.api.v1.SelfService/ChangeEmail"
const OperationSelfServiceInfo = "/goddess.api.v1.SelfService/Info"
const OperationSelfServiceNamespaces = "/goddess.api.v1.SelfService/Namespaces"
const OperationSelfServiceRefreshToken = "/goddess.api.v1.SelfService/RefreshToken"

type SelfServiceHTTPServer interface {
	ChangeAvatar(context.Context, *ChangeAvatarRequest) (*ChangeAvatarReply, error)
	ChangeEmail(context.Context, *ChangeEmailRequest) (*ChangeEmailReply, error)
	Info(context.Context, *InfoRequest) (*UserItem, error)
	Namespaces(context.Context, *InfoRequest) (*NamespacesReply, error)
	RefreshToken(context.Context, *RefreshTokenRequest) (*RefreshTokenReply, error)
}

func RegisterSelfServiceHTTPServer(s *http.Server, srv SelfServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/self/info", _SelfService_Info0_HTTP_Handler(srv))
	r.GET("/v1/self/namespaces", _SelfService_Namespaces0_HTTP_Handler(srv))
	r.PUT("/v1/self/change-email", _SelfService_ChangeEmail0_HTTP_Handler(srv))
	r.PUT("/v1/self/change-avatar", _SelfService_ChangeAvatar0_HTTP_Handler(srv))
	r.GET("/v1/self/refresh-token", _SelfService_RefreshToken0_HTTP_Handler(srv))
}

func _SelfService_Info0_HTTP_Handler(srv SelfServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in InfoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSelfServiceInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Info(ctx, req.(*InfoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UserItem)
		return ctx.Result(200, reply)
	}
}

func _SelfService_Namespaces0_HTTP_Handler(srv SelfServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in InfoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSelfServiceNamespaces)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Namespaces(ctx, req.(*InfoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*NamespacesReply)
		return ctx.Result(200, reply)
	}
}

func _SelfService_ChangeEmail0_HTTP_Handler(srv SelfServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ChangeEmailRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSelfServiceChangeEmail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ChangeEmail(ctx, req.(*ChangeEmailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ChangeEmailReply)
		return ctx.Result(200, reply)
	}
}

func _SelfService_ChangeAvatar0_HTTP_Handler(srv SelfServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ChangeAvatarRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSelfServiceChangeAvatar)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ChangeAvatar(ctx, req.(*ChangeAvatarRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ChangeAvatarReply)
		return ctx.Result(200, reply)
	}
}

func _SelfService_RefreshToken0_HTTP_Handler(srv SelfServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RefreshTokenRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSelfServiceRefreshToken)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RefreshToken(ctx, req.(*RefreshTokenRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RefreshTokenReply)
		return ctx.Result(200, reply)
	}
}

type SelfServiceHTTPClient interface {
	ChangeAvatar(ctx context.Context, req *ChangeAvatarRequest, opts ...http.CallOption) (rsp *ChangeAvatarReply, err error)
	ChangeEmail(ctx context.Context, req *ChangeEmailRequest, opts ...http.CallOption) (rsp *ChangeEmailReply, err error)
	Info(ctx context.Context, req *InfoRequest, opts ...http.CallOption) (rsp *UserItem, err error)
	Namespaces(ctx context.Context, req *InfoRequest, opts ...http.CallOption) (rsp *NamespacesReply, err error)
	RefreshToken(ctx context.Context, req *RefreshTokenRequest, opts ...http.CallOption) (rsp *RefreshTokenReply, err error)
}

type SelfServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewSelfServiceHTTPClient(client *http.Client) SelfServiceHTTPClient {
	return &SelfServiceHTTPClientImpl{client}
}

func (c *SelfServiceHTTPClientImpl) ChangeAvatar(ctx context.Context, in *ChangeAvatarRequest, opts ...http.CallOption) (*ChangeAvatarReply, error) {
	var out ChangeAvatarReply
	pattern := "/v1/self/change-avatar"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSelfServiceChangeAvatar))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SelfServiceHTTPClientImpl) ChangeEmail(ctx context.Context, in *ChangeEmailRequest, opts ...http.CallOption) (*ChangeEmailReply, error) {
	var out ChangeEmailReply
	pattern := "/v1/self/change-email"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSelfServiceChangeEmail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SelfServiceHTTPClientImpl) Info(ctx context.Context, in *InfoRequest, opts ...http.CallOption) (*UserItem, error) {
	var out UserItem
	pattern := "/v1/self/info"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSelfServiceInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SelfServiceHTTPClientImpl) Namespaces(ctx context.Context, in *InfoRequest, opts ...http.CallOption) (*NamespacesReply, error) {
	var out NamespacesReply
	pattern := "/v1/self/namespaces"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSelfServiceNamespaces))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SelfServiceHTTPClientImpl) RefreshToken(ctx context.Context, in *RefreshTokenRequest, opts ...http.CallOption) (*RefreshTokenReply, error) {
	var out RefreshTokenReply
	pattern := "/v1/self/refresh-token"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSelfServiceRefreshToken))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}